scripts/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
opencv-python==4.8.1.78
face-alignment==1.3.5
scikit-image==0.21.0
PyYAML==6.0.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
